Extract the Collateral Exposure snapshot into a partial
Increment 2j of the endpoint-monolith decomposition. Moves the collateral
exposure snapshot builders (BuildCollateralExposureSnapshot and its
BuildFallbackCollateralRows helper) out of
src/Meridian.Ui.Shared/Endpoints/WorkstationEndpoints.cs into a new sibling
partial WorkstationEndpoints.CollateralExposure.cs.

Behavior-preserving verbatim relocation: WorkstationEndpoints is one partial
class, so the inline GetWorkstationCollateralExposure route lambda keeps calling
BuildCollateralExposureSnapshot unchanged, and the moved builder still reaches
the core-resident NormalizeOperatorInboxToken helper across the partial. Closed
call graph verified: both methods have zero references in other partials; the
only entry is the single inline route registration in core.

Left in core (shared): the collateral route lambda and NormalizeOperatorInboxToken
(shared with the Operator Inbox partial).

Core WorkstationEndpoints.cs drops 5,636 -> 5,567 lines; ratchet baseline
retightened.
